+++ src/Dose/Execute.cs
@@ -1,16 +1,12 @@
 ﻿/* PROJECT: Dose (https://github.com/aprettycoolprogram/Dose)
  *    FILE: Dose.Execute.cs
- * UPDATED: 7-1-2021-6:56 PM
+ * UPDATED: 7-1-2021-8:40 PM
  * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
  *          Copyright 2021 A Pretty Cool Program All rights reserved
  */
 
-using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
 using NTST.ScriptLinkService.Objects;
 
 namespace Dose
@@ -20,7 +16,7 @@ namespace Dose
         public static Dictionary<string, string> DoseSetting;
 
         /// <summary>
-        /// Executes a MAWS action for the InptAdmitDate command.
+        /// Executes a MAWS action for the Dose command.
         /// </summary>
         /// <param name="sentOptionObject2015">The original OptionObject2015 sent from myAvatar.</param>
         /// <param name="mawsRequest">         The MAWS Request string.</param>
@@ -29,27 +25,27 @@ namespace Dose
         {
             DoseSetting = Settings.GetSettings();
 
-            var mawsAction                = RequestSyntaxEngine.RequestComponent.GetAction(mawsRequest);
-            var mawsOption                = RequestSyntaxEngine.RequestComponent.GetOption(mawsRequest);
-            Logger.Timestamped.LogEvent(DoseSetting["Logging"].ToLower(), "TRACE", Assembly.GetExecutingAssembly().GetName().Name, $"Execute InptAdmitDate Action: {mawsAction} [Option={mawsOption}]");
+            var mawsAction = RequestSyntaxEngine.RequestComponent.GetAction(mawsRequest);
+            var mawsOption = RequestSyntaxEngine.RequestComponent.GetOption(mawsRequest);
+            Logger.Timestamped.LogEvent(DoseSetting["Logging"].ToLower(), "TRACE", Assembly.GetExecutingAssembly().GetName().Name, $"Execute Dose Action: {mawsAction} [Option={mawsOption}]");
 
-            var inptAdmitDateOptionObject = new OptionObject2015();
+            var doseOptionObject = new OptionObject2015();
 
             switch(mawsAction)
             {
                 case "verifypercentage":
-                    Logger.Timestamped.LogEvent(DoseSetting["Logging"].ToLower(), "TRACE", Assembly.GetExecutingAssembly().GetName().Name, $"Executing InptAdmitDate Action: comparepreadmittoadmit [Option={mawsOption}]");
-                    inptAdmitDateOptionObject = mawsOption == "testing"
+                    Logger.Timestamped.LogEvent(DoseSetting["Logging"].ToLower(), "TRACE", Assembly.GetExecutingAssembly().GetName().Name, $"Executing Dose Action: comparepreadmittoadmit [Option={mawsOption}]");
+                    doseOptionObject = mawsOption == "testing"
                         ? Verify.Percentage_Testing(sentOptionObject2015, DoseSetting)
                         : Verify.Percentage(sentOptionObject2015, DoseSetting);
                     break;
 
                 default:
-                    Logger.Timestamped.LogEvent(DoseSetting["Logging"].ToLower(), "TRACE", Assembly.GetExecutingAssembly().GetName().Name, $"Invalid InptAdmitDate Action: \"{mawsAction}\"");
+                    Logger.Timestamped.LogEvent(DoseSetting["Logging"].ToLower(), "ERROR", Assembly.GetExecutingAssembly().GetName().Name, $"Invalid Dose Action: \"{mawsAction}\"");
                     break;
             }
 
-            return inptAdmitDateOptionObject;
+            return doseOptionObject;
         }
     }
 }