+++ src/Dose/Settings.cs
@@ -1,53 +1,53 @@
 ﻿/* PROJECT: Dose (https://github.com/aprettycoolprogram/Dose)
  *    FILE: Dose.Settings.cs
- * UPDATED: 7-1-2021-6:56 PM
+ * UPDATED: 7-1-2021-8:42 PM
  * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
  *          Copyright 2021 A Pretty Cool Program All rights reserved
  */
 
-using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Dose
 {
     public class Settings
     {
+        /// <summary>
+        /// Gets the settings information from Dose.settings.
+        /// </summary>
+        /// <returns>The MAWS settings dictionary.</returns>
         public static Dictionary<string, string> GetSettings()
         {
-            var settings = new Dictionary<string, string>();
+            var doseSettings     = new Dictionary<string, string>();
+            var doseSettingsPath = @"C:/MAWS/MAWS.settings";
 
-            if(File.Exists(@"C:/MAWS/Dose.settings"))
+            if(File.Exists(doseSettingsPath))
             {
-                var settingsFileStream = new System.IO.StreamReader(@"C:/MAWS/Dose.settings");
-
-                var settingsAsList = new List<string>();
-                var line = "";
+                var    settingsFileStream = new StreamReader(doseSettingsPath);
+                var    settingsAsList     = new List<string>();
+                string settingLine;
 
                 using(settingsFileStream)
                 {
-                    while((line = settingsFileStream.ReadLine()) != null)
+                    while((settingLine = settingsFileStream.ReadLine()) != null)
                     {
-                        settingsAsList.Add(line.Trim());
+                        settingsAsList.Add(settingLine.Trim());
                     }
                 }
 
-                var keyValuePair  = new string[2];
+                string[] keyValuePair;
 
                 foreach(var item in settingsAsList)
                 {
                     if(!item.StartsWith("#"))
                     {
                         keyValuePair = item.Split('=');
-                        settings.Add(keyValuePair[0], keyValuePair[1]);
+                        doseSettings.Add(keyValuePair[0], keyValuePair[1]);
                     }
                 }
             }
 
-            return settings;
+            return doseSettings;
         }
     }
 }