+++ src/Dose/Verify.cs
@@ -1,20 +1,16 @@
 ﻿/* PROJECT: Dose (https://github.com/aprettycoolprogram/Dose)
  *    FILE: Dose.Verify.cs
- * UPDATED: 7-1-2021-6:56 PM
+ * UPDATED: 7-1-2021-8:44 PM
  * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
  *          Copyright 2021 A Pretty Cool Program All rights reserved
  */
 
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using NTST.ScriptLinkService.Objects;
 
 namespace Dose
 {
-    class Verify
+    internal class Verify
     {
         public static OptionObject2015 Percentage(OptionObject2015 sentOptionObject, Dictionary<string, string> doseSetting)
         {