+++ src/Logger/Maintenance.cs
@@ -1,8 +1,13 @@
 ﻿/* PROJECT: Logger (https://github.com/aprettycoolprogram/Logger)
  *    FILE: Logger.Maintenance.cs
- * UPDATED: 7-1-2021-12:39 PM
+ * UPDATED: 7-1-2021-8:45 PM
  * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
  *          Copyright 2021 A Pretty Cool Program All rights reserved
+ */
+
+/* Performs logging maintenance.
+ *
+ * Development notes/comments can be found at the end of this class.
  */
 
 using System.IO;
@@ -13,7 +18,7 @@ namespace Logger
     public class Maintenance
     {
         /// <summary>
-        /// Confirm existance of and/or create the log directory.
+        /// Confirm existance of and/or create a log directory.
         /// </summary>
         public static void ConfirmLogDirectoryExists(string directoryPath)
         {