+++ src/NewDevelopment/Properties/AssemblyInfo.cs
@@ -1,22 +1,15 @@
-﻿/* PROJECT: InptAdmitDate (https://github.com/aprettycoolprogram/InptAdmitDate)
- *    FILE: %Namespace%.AssemblyInfo.cs
- * UPDATED: 7-1-2021-7:01 PM
- * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
- *          Copyright 2021 A Pretty Cool Program All rights reserved
- */
-
-using System.Reflection;
+﻿using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("InptAdmitDate")]
+[assembly: AssemblyTitle("NewDevelopment")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("InptAdmitDate")]
+[assembly: AssemblyProduct("NewDevelopment")]
 [assembly: AssemblyCopyright("Copyright ©  2021")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
@@ -27,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("566c8ed3-5fef-4dcc-8b33-b2090e7727f7")]
+[assembly: Guid("f1b45ad4-e6ea-47cb-9ad7-242f8d49a2f1")]
 
 // Version information for an assembly consists of the following four values:
 //