+++ src/RequestSyntaxEngine/TestFunctionality.cs
@@ -1,6 +1,6 @@
 ﻿/* PROJECT: RequestSyntaxEngine (https://github.com/aprettycoolprogram/RequestSyntaxEngine)
  *    FILE: RequestSyntaxEngine.TestFunctionality.cs
- * UPDATED: 7-1-2021-11:14 AM
+ * UPDATED: 7-1-2021-8:46 PM
  * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
  *          Copyright 2021 A Pretty Cool Program All rights reserved
  */
@@ -23,7 +23,8 @@ namespace RequestSyntaxEngine
                               $" MAWS Action: {RequestComponent.GetAction(mawsRequest)}{Environment.NewLine}" +
                               $" MAWS Option: {RequestComponent.GetOption(mawsRequest)}";
 
-            Logger.Timestamped.LogEvent("testing", "TESTING", Assembly.GetExecutingAssembly().GetName().Name, logMessage);
+            // The RequestSyntaxEngine doesn't have it's own settings file, so we hardcode the logSetting parameter.
+            Logger.Timestamped.LogEvent("trace", "TRACE", Assembly.GetExecutingAssembly().GetName().Name, logMessage);
         }
     }
 }